requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# Smart Contract Summary & Q&A Assistant – Python Dependencies
# ─────────────────────────────────────────────────────────────────────────────

# --- Web / API Framework ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9          # FastAPI file-upload support

# --- LangChain core + community ---
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# --- Vector Store (FAISS – pure C++/numpy, no Pydantic dependency) ---
faiss-cpu>=1.7.4

# --- Embeddings (local, no API key needed) ---
sentence-transformers>=3.0.0

# --- LLM  (choose ONE of the two blocks below)
# Option A – OpenAI GPT  (requires OPENAI_API_KEY env var)
openai>=1.30.0
langchain-openai>=0.1.0

# Option B – Local HuggingFace model (heavier, no API key needed)
# Uncomment the line below if you want a fully offline experience
# transformers>=4.40.0

# --- Document Parsing ---
PyMuPDF>=1.24.0        # fitz – fast PDF parsing
pdfplumber>=0.11.0     # alternate PDF parser (tables, layout)
python-docx>=1.1.0     # DOCX support

# --- Frontend UI ---
gradio>=4.36.0

# --- Utilities ---
python-dotenv>=1.0.0   # load .env files
tiktoken>=0.7.0        # token counting (needed by LangChain / OpenAI)
tqdm>=4.66.0           # progress bars

# --- Testing (optional) ---
pytest>=8.2.0
httpx>=0.27.0          # async HTTP client used by FastAPI test client
